Final Project/requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
pyenchant==3.2.2
Pygments==2.17.2
rich==13.7.1
